Review feedback: extract interfaces to a separate folder
