+++ src/main/reactjs/src/interfaces/translator.ts
@@ -1,0 +1,10 @@
+interface LanguagePair {
+  from: string;
+  to: string;
+}
+
+export interface TranslatorDetails {
+  name: string;
+  languagePairs: Array<LanguagePair>;
+  areasOfOperation: Array<string>;
+}
